move config to internal
