+++ .gitignore
@@ -1,3 +1,4 @@
 .idea/
 **/.DS_Store
 **/config.yml
+**/config.yaml
