+++ .vscode/settings.json
@@ -1,0 +1,7 @@
+{
+  "[go]": {
+    "editor.defaultFormatter": "golang.go"
+  },
+  "editor.formatOnPaste": true,
+  "go.formatTool": "goimports",
+}
