+++ payment/internal/config/config.example.yml
@@ -1,0 +1,8 @@
+server:
+  port: 3001
+log:
+  level: info
+  service: payment
+omise:
+  public_key: omisepublickey
+  secret_key: omisesecretkey
